// backend/src/models/Jogo.ts
export class Jogo{
    id: number = 0;
    nome: string = '';
    data_lancamento: string = '';
    descricao: string = '';
    idioma: string[] = [];
    nota: number = 0;
    genero: string[] = [];
    imagem: string = 'https://images.igdb.com/igdb/image/upload/t_cover_big/';
    produtora: string[] = [];
    plataformas: string[] = [];
}